Terminado el E28 empezado el E29

+++ Assets/Scenes/E17/SCRIPTS/Diag_Mov_Speedable.cs
@@ -4,7 +4,7 @@ using static UnityEngine.RuleTile.TilingRuleOutput;
 public class Diag_Mov_Speedable : MonoBehaviour
 {
     private Vector3 Dir = new Vector3(1, -1, 1);
-    public int SPEED = 0;
+    public float SPEED = 0;
 
     // Start is called once before the first execution of Update after the MonoBehaviour is created
     void Start()
